Kun admin kan slette konto og opdater konto

+++ app/frontend/pages/profile.php
@@ -1,3 +1,14 @@
+<?php
+require_once 'app/backend/core/Init.php';
+checkRole('administrator');
+function checkRole($requiredRole) {
+    $user = new User();
+    if ($user->data()->role !== $requiredRole) {
+        return false;
+    }
+    return true;
+}
+?>
 <!DOCTYPE html>
 <html lang="en">
 <head>
@@ -39,12 +50,15 @@
                         </tbody>
                     </table>
                     <div class="text-right">
-                        <a href="update-account.php" class="btn btn-primary custom-btn custom-btn-primary" aria-label="Opdater brugeroplysninger" title="Opdater brugeroplysninger">
-                            <i class="bi bi-person-up"></i>
-                        </a>
-                        <a href="delete-account.php" class="btn btn-danger custom-btn custom-btn-danger" aria-label="Slet brugerkonto" title="Slet brugerkonto"> 
-                            <i class="bi bi-trash3"></i>
-                        </a>
+<?php if (checkRole('administrator')): ?>
+                    <a href="update-account.php" class="btn btn-primary custom-btn custom-btn-primary" aria-label="Opdater brugeroplysninger" title="Opdater brugeroplysninger"> 
+                    <i class="bi bi-person-up"></i>
+                    </a>
+                    <a href="delete-account.php" class="btn btn-danger custom-btn custom-btn-danger" aria-label="Slet brugerkonto" title="Slet brugerkonto"> 
+                <i class="bi bi-trash3"></i>
+                </a>
+                    
+<?php endif; ?>
                     </div>
                 </div>
             </div>
